# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rara Ads Api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "Localhost"
        },
        {
            "url": "https://raraads.bidhee.net/api/",
            "description": "Raraads OpenApi Server"
        }
    ],
    "paths": {
...
        "/driver/update-password": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "Change Password",
                "description": "Change Password",
                "operationId": "changePassword",
                "parameters": [
                    {
                        "name": "old_password",
                        "in": "query",
                        "description": "Old Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password Confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "201": {
                        "description": "Object Created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthorized Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}